llamaindex/dataqa-trulens/requirements.txt
openai
llama_index
trulens-eval
python-dotenv
utils
pypdf
transformers
torch
langchain
langchain-community